requirements.txt
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
Pillow==9.4.0
protobuf==3.20.3
